Create install_package.yml
This play will check if software nano and vim are installed on machines and install them if they are not already there.  This works for RedHat and  Debian based distributions, but could be extended.  You can also replace the defined hosts and software with "{{ HOST }}" and "{{ SOFTWARE }}" so any software can be searched for on any machine.